% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-exceed.R
\name{exceed_dist}
\alias{exceed_dist}
\title{Construct a distribution for exceedance probability}
\usage{
exceed_dist(
  pfun,
  params = function(estimate, error) list(),
  lower.tail = FALSE
)
}
\arguments{
\item{pfun}{A CDF function. If \code{NULL} (the default), \code{stats::pnorm()} is used.}

\item{params}{A function that takes \code{estimate} and \code{error} as inputs
and returns a list of parameters required by \code{pfun}.}

\item{lower.tail}{Logical. If \code{FALSE} (the default), the constructed
function returns \code{P(X > q)}; if \code{TRUE}, it returns \code{P(X <= q)}.}
}
\description{
Build a function \code{dist_fun(q, estimate, error)} that returns the exceedance
probability \code{P(X > q)} by combining a base R cumulative distribution
function \code{pfun} with a user-defined parameter transformation based on
\code{estimate} and \code{error}.
}
\examples{
data(nc)
dist_exp <- exceed_dist(stats::pexp, function(estimate, error)
  list(rate = 1 / estimate))
ex_pr <- dist_exp(q = 3, estimate = nc$v1, error = nc$v2)

}
\seealso{
\code{\link[=scale_fill_exceed]{scale_fill_exceed()}} for use in exceedance probability maps.
}
