% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-exceed.R
\name{geom_sf_exceed}
\alias{geom_sf_exceed}
\title{Geom layer function for exceedance probability map on sf objects}
\usage{
geom_sf_exceed(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  dist_fun = dist_norm,
  threshold = 0,
  palette = "YlOrRd",
  direction = 1,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping that must include \code{geometry}, \code{estimate}, and \code{error}.}

\item{data}{An sf object.}

\item{geom}{Geom type to use.}

\item{position}{Position adjustment.}

\item{show.legend}{Logical; whether to display a legend.}

\item{inherit.aes}{Logical; whether to inherit global aesthetics.}

\item{dist_fun}{Function to compute upper tail probabilities that takes the
argument the \code{threshold}, \code{v1} and \code{v2}.}

\item{threshold}{Numeric threshold for exceedance probability.}

\item{palette}{Palette name for \code{scale_fill_distiller}.}

\item{direction}{Direction of the palette (1 or -1).}

\item{...}{Additional parameters passed to \code{layer_sf()}.}
}
\value{
A list containing a ggproto object drawing sf geometries colored by
exceedance probability and a ggplot2 scale object for the fill aesthetic.
}
\description{
geom_sf_exceed applies StatExceed to sf data and returns a map that visualises
the probability of exceeding some nominated threshold of concern.
}
\examples{
set.seed(10086)
data(nc)
ggplot(nc) +
  geom_sf_exceed(
    mapping = aes(
      v1 = value,
      v2 = sd,
      fill = after_stat(pr_exc)
    ),
    threshold = 0.5,
    dist_fun  = stats::pnorm
  ) +
  coord_sf()

}
