% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-glyph.R
\docType{data}
\name{StatGlyph}
\alias{StatGlyph}
\alias{geom_sf_glyph}
\title{Generate glyph maps on sf objects}
\format{
An object of class \code{StatGlyph} (inherits from \code{StatSf}, \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatGlyph

geom_sf_glyph(
  mapping = NULL,
  data = NULL,
  size = 70,
  style = "icone",
  max_v2 = NULL,
  position = "identity",
  show.legend = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}.
\code{v1} and \code{v2} are required, which are the variables used for glyph fill
and rotation, respectively.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{size}{A positive numeric scaling factor controlling glyph size.
Larger values produce smaller glyphs.}

\item{style}{Either \code{"icone"} or \code{"semi"}. Controls the glyph shape.}

\item{max_v2}{Numeric value setting the upper limit for \code{v2}.}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.

You can also set this to one of "polygon", "line", and "point" to
override the default legend.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}
}
\value{
A list of ggplot2 layer objects.
}
\description{
\code{geom_sf_glyph()} adds a glyph map layer based on simple feature (sf) objects
to a ggplot. A glyph map is essentially a centroid-based map, where each
region is represented by a rotated glyph, and the rotation angle indicates
the value of \code{v2} specified in the mapping.
}
\section{Glyph map layer contents}{

The layer returned by \code{geom_sf_glyph()} actually contains two scales,
corresponding to the two variables specified in the mapping.
Therefore, modifying the scale for \code{v1} will trigger a warning
indicating that the scale for \code{fill} is being replaced.
}

\examples{
data(nc)

# Basic glyph map
p <- ggplot(nc) + geom_sf_glyph(mapping = aes(v1 = value, v2 = sd))
p1 <- ggplot(nc) + geom_sf_glyph(mapping = aes(v1 = value, v2 = sd), style = "semi")

# Customize labels and theme
p + labs(title = "glyph map on nc") + theme(legend.position = "left", legend.box = "horizontal")

# Replacing the internal fill scale triggers a message
# ("Scale for fill is already present. Adding another scale for fill...")
p + scale_fill_distiller(palette = "Blues")

}
\keyword{datasets}
