% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-bivariate.R
\name{GuideBivariate}
\alias{GuideBivariate}
\alias{guide_bivariate}
\title{Bivariate colour guide}
\usage{
GuideBivariate

guide_bivariate(
  theme = NULL,
  title = NULL,
  order = 0,
  position = NULL,
  rotated = FALSE,
  angle = 45
)
}
\arguments{
\item{theme}{A \code{\link[ggplot2:theme]{theme}} object to style the guide individually or
differently from the plot's theme settings. The \code{theme} argument in the
guide partially overrides, and is combined with, the plot's theme.
Arguments that apply to a single legend are respected, most of which have
the \code{legend}-prefix. Arguments that apply to combined legends
(the legend box) are ignored, including \code{legend.position},
\verb{legend.justification.*}, \code{legend.location} and \verb{legend.box.*}.}

\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{order}{positive integer less than 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{position}{A character string indicating where the legend should be
placed relative to the plot panels.
One of "top", "right", "bottom", "left", or "inside".}

\item{rotated}{Logical indicating whether the guide should be displayed in a
rotated orientation.}

\item{angle}{A numeric value specifying the rotation angle of the guide.}
}
\description{
The bivariate colour guide is displayed as a grid, where each cell represents
the mapping between a colour and the corresponding bin intervals of the two
variables along the guide axes.
}
