% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc}
\alias{nc}
\title{North Carolina SIDS data}
\format{
A \code{sf} object.
}
\usage{
nc
}
\description{
The dataset \code{nc} is derived from the North Carolina shapefile (\code{nc.shp})
included in the \pkg{sf} package. Two random variables, \code{value} and \code{sd},
have been added for demonstration purposes.

Further details about the original data can be found in the
\href{https://r-spatial.github.io/spdep/articles/sids.html}{spdep package vignette}.
}
\examples{
head(nc)

plot(sf::st_geometry(nc))
}
\keyword{datasets}
