% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-exceed.R
\docType{data}
\name{ScaleExceed}
\alias{ScaleExceed}
\alias{scale_fill_exceed}
\alias{scale_colour_exceed}
\title{#' Construct a distribution for exceedance probability
#'
#' Build a function \code{dist_fun(q, estimate, error)} that returns the exceedance
#' probability \code{P(X > q)} by combining a base R cumulative distribution
#' function \code{pfun} with a user-defined parameter transformation based on
#' \code{estimate} and \code{error}.
#'
#' @param pfun A CDF function. If \code{NULL} (the default), \code{stats::pnorm()} is used.
#' @param params A function that takes \code{estimate} and \code{error} as inputs
#'   and returns a list of parameters required by \code{pfun}.
#' @param lower.tail Logical. If \code{FALSE} (the default), the constructed
#'   function returns \code{P(X > q)}; if \code{TRUE}, it returns \code{P(X <= q)}.
#'
#' @examples
#' data(nc)
#' dist_exp <- exceed_dist(stats::pexp, function(estimate, error)
#'   list(rate = 1 / estimate))
#' ex_pr <- dist_exp(q = 3, estimate = nc$v1, error = nc$v2)
#'
#' @seealso \code{\link[=scale_fill_exceed]{scale_fill_exceed()}} for use in exceedance probability maps.
#' @export
exceed_dist <- function(pfun,
params = function(estimate, error)
list(),
lower.tail = FALSE) {
browser()
stopifnot(is.function(pfun), is.function(params))
function(q, estimate, error) {
browser()
args <- c(list(q = q, lower.tail = lower.tail), params(estimate, error))
do.call(pfun, args)
}
}}
\format{
An object of class \code{ScaleExceed} (inherits from \code{ScaleContinuous}, \code{Scale}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
ScaleExceed

scale_fill_exceed(
  name = NULL,
  palette = "Oranges",
  type = "seq",
  direction = 1,
  dist_fun = NULL,
  threshold,
  limits = c(0, 1),
  na.value = NA,
  guide = "colourbar",
  aesthetics = "fill",
  ...
)

scale_colour_exceed(
  name = NULL,
  palette = "Oranges",
  type = "seq",
  direction = 1,
  dist_fun = NULL,
  threshold,
  limits = c(0, 1),
  na.value = NA,
  guide = "colourbar",
  aesthetics = "colour",
  ...
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{palette}{A palette function that when called with a numeric vector with
values between 0 and 1 returns the corresponding output values
(e.g., \code{\link[scales:pal_area]{scales::pal_area()}}).}

\item{type}{One of "seq" (sequential), "div" (diverging) or "qual"
(qualitative)}

\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are as output by \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}}. If -1, the
order of colours is reversed.}

\item{dist_fun}{A function used to compute the exceedance probability.
Typically, this should be the output of \code{\link[=exceed_dist]{exceed_dist()}}. If \code{NULL}
(the default), a normal distribution with \code{stats::pnorm()} is used.}

\item{threshold}{A numeric value specifying the threshold \code{q} in the exceedance
probability expression \code{P(X > q)}.}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
Note that setting limits on positional scales will \strong{remove} data outside of the limits.
If the purpose is to zoom, use the limit argument in the coordinate system
(see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}).
}}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\description{
\verb{scale_*_exceed} computes exceedance probabilities from a specified
distribution and maps onto a continuous gradient colour scale.
}
\examples{
ggplot(nc) +
  geom_sf(aes(fill = duo_exceed(value, sd))) +
  scale_fill_exceed(threshold = 1.64)

}
\keyword{datasets}
