% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-fill-bivariate.R
\name{scale_fill_bivariate}
\alias{scale_fill_bivariate}
\title{Bivariate fill scale for bivariate map}
\usage{
scale_fill_bivariate(
  colrange,
  subtractive = FALSE,
  flip_vertical = FALSE,
  flip_horizontal = FALSE,
  ...
)
}
\arguments{
\item{colrange}{List with a character vector of length two called colour and a numeric
vector of length two called difC. colour contains two colour names
from the colours() range or valid hexadecimal colors. difC contains two
integers (1, 2, 3 or 4), which control how much a colour changes in value across the grid.}

\item{subtractive}{A logical evaluating to TRUE or FALSE indicating whether the colour mixing is subtractive or additive.}

\item{...}{Additional arguments passed to \code{scale_fill_manual()}.}

\item{flipVertical}{Whether the palette should be flipped vertically (ie. replace top portion with bottom portion).}

\item{flipHorizontal}{Whether the palette should be flipped horizontally (ie. replace left portion with right portion).}
}
\value{
A ggplot2 scale object that maps fill to the 9 blended colors.
}
\description{
Create a 3×3 bivariate color scale by blending two directional color ramps.
}
