% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-bivariate.R
\docType{data}
\name{ScaleBivariate}
\alias{ScaleBivariate}
\title{ScaleBivariate}
\format{
An object of class \code{ScaleBivariate} (inherits from \code{ScaleDiscrete}, \code{Scale}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
ScaleBivariate
}
\description{
All \verb{scale_*_bivariate()} functions return a \code{ScaleBivariate} ggproto object.
This object defines how bivariate scales map data values to visual aesthetics.
}
\seealso{
\link[ggplot2:ggplot2-ggproto]{ggplot2::Scale} for the base \code{ggproto} class that all scale objects inherit from.
}
\keyword{datasets}
