% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-glyph.R
\docType{data}
\name{StatGlyph}
\alias{StatGlyph}
\title{Geom layer function for glyph map on sf objects}
\format{
An object of class \code{StatGlyph} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatGlyph
}
\arguments{
\item{mapping}{Aesthetic mapping that must include \code{geometry}, \code{estimate}, and \code{error}.}

\item{data}{An sf object.}

\item{geom}{Geom type, default "polygon"}

\item{position}{Position adjustment.}

\item{show.legend}{Logical; whether to display a legend.}

\item{inherit.aes}{Logical; whether to inherit global aesthetics.}

\item{size}{Glyph scale factor.}

\item{glyph}{Glyph type: "icone" or "semi".}

\item{max_error}{Maximum error scale.}

\item{...}{Other params passed to layer()}
}
\description{
\code{geom_sf_glyph} applies \code{StatGlyph} to sf data and returns a centroid map
with color and rotation based on estimate and error values.
}
\examples{
set.seed(10086)

nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE) \%>\%
  mutate(value = rnorm(n()), sd = rnorm(n()))

ggplot(nc, aes(geometry = geometry, estimate = value, error = sd)) +
  geom_sf_glyph(size = 50, glyph = "icone") +
  coord_sf() + theme_void()

}
\keyword{datasets}
