% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-pixel.R
\docType{data}
\name{StatPixel}
\alias{StatPixel}
\title{Geom layer function for pixel map on sf objects}
\format{
An object of class \code{StatPixel} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatPixel
}
\arguments{
\item{mapping}{Aesthetic mapping that must include \code{geometry}, \code{estimate}, \code{error} and \code{id}.}

\item{data}{An sf object.}

\item{pixelSize}{Number of grid cells along each axis.}

\item{distribution}{Sampling distribution: "uniform", "normal", or "discrete".}

\item{id_col}{Name of the ID column.}

\item{show.legend}{Logical; show legend.}

\item{inherit.aes}{Logical; inherit global aesthetics.}

\item{...}{Other parameters passed to \code{layer()}.}
}
\value{
A polygon layer showing pixel values.
}
\description{
\code{geom_sf_pixel} applies \code{StatPixel} to sf data, pixelate each region and
sample each pixel's value from a specified distribution.
}
\examples{
set.seed(10086)
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE) \%>\%
  mutate(value = rnorm(n()), sd = rnorm(n()))

names(nc)[ names(nc) == "NAME" ] <- "id"
names(nc)[ names(nc) == "value" ] <- "estimate"
names(nc)[ names(nc) == "sd" ] <- "error"

ggplot() +
  geom_sf_pixel(
    data = nc,
    mapping = aes(geometry = geometry),
    id_col = "id",
    pixelSize = 40,
    distribution = "uniform"
  ) +
  scale_fill_distiller(palette = "Blues", direction = 1) +
  geom_sf(data = nc, fill = NA, color = 'black', size = 0.2)

}
\keyword{datasets}
