% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsup-quantize.R
\name{vsup_quantize}
\alias{vsup_quantize}
\title{Tree quantization for VSUPs}
\usage{
vsup_quantize(
  v,
  u,
  layers = 4,
  branch = 2L,
  breaks = list(NULL, NULL),
  limits = list(NULL, NULL),
  transform = list("identity", "identity")
)
}
\arguments{
\item{v}{Numeric vector of value variable.}

\item{u}{Numeric vector of uncertainty variable.}

\item{layers}{An integer specifying the number of uncertainty levels.}

\item{branch}{An integer specifying the branching factor used to allocate
value bins across uncertainty levels. The maximum number of value bins is
\code{branch^(layers - 1)}, and higher uncertainty levels are assigned fewer
value bins.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{transform}{A list of one or two transformations applied to continuous
variables before binning. Each element can be a transformation name or a
transformer object accepted by \code{\link[scales:as.transform]{scales::as.transform()}}.}
}
\value{
A list containing quantized leaf ids and break information.
}
\description{
Quantize value and uncertainty variables into a hierarchical VSUP tree.
}
