% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-bivariate.R
\docType{data}
\name{ScaleBivariate}
\alias{ScaleBivariate}
\alias{bivariate_scale}
\alias{scale_fill_bivariate}
\alias{scale_color_bivariate}
\alias{scale_colour_bivariate}
\title{Bivariate colour scale constructor}
\format{
An object of class \code{ScaleBivariate} (inherits from \code{ScaleDiscrete}, \code{Scale}, \code{ggproto}, \code{gg}) of length 17.
}
\usage{
ScaleBivariate

bivariate_scale(
  aesthetics,
  ...,
  name = waiver(),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  na.value = NA,
  na.translate = TRUE,
  drop = FALSE,
  guide = waiver(),
  colours = c("gold", "red4"),
  palette_fun = NULL,
  palette_params = list(),
  n_breaks = c(4, 4),
  bin_method = c("equal", "equal"),
  var1_name = NULL,
  var2_name = NULL,
  super = ScaleBivariate
)

scale_fill_bivariate(
  ...,
  name = waiver(),
  var1_name = NULL,
  var2_name = NULL,
  colours = c("gold", "red4"),
  palette_fun = NULL,
  palette_params = list(),
  n_breaks = c(4, 4),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  bin_method = c("equal", "equal"),
  na.value = NA,
  aesthetics = "fill",
  guide = guide_bivariate()
)

scale_color_bivariate(
  ...,
  name = waiver(),
  var1_name = NULL,
  var2_name = NULL,
  colours = c("gold", "red4"),
  palette_fun = NULL,
  palette_params = list(),
  n_breaks = c(4, 4),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  bin_method = c("equal", "equal"),
  na.value = NA,
  aesthetics = "colour",
  guide = guide_bivariate()
)

scale_colour_bivariate(
  ...,
  name = waiver(),
  var1_name = NULL,
  var2_name = NULL,
  colours = c("gold", "red4"),
  palette_fun = NULL,
  palette_params = list(),
  n_breaks = c(4, 4),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  bin_method = c("equal", "equal"),
  na.value = NA,
  aesthetics = "colour",
  guide = guide_bivariate()
)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{...}{Other arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}

\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{A list of two numeric vectors specifying bin boundaries for
each variable. If \code{waiver()}, breaks are computed from the data according
to \code{n_breaks} and \code{bin_method}.}

\item{labels}{A list of two character vectors or labelling functions
used to label the bin boundaries for each variable. If \code{waiver()},
default numeric labels are used.}

\item{limits}{A list of two numeric vectors specifying the range of
values to include for each variable.}

\item{transform}{A list of two transformations applied to the variables
before binning. Each element can be a transformation name or a
transformer object accepted by \code{\link[scales:new_transform]{scales::as.transform()}}.}

\item{na.value}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}

\item{na.translate}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}

\item{drop}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} includes the levels in the factor. Please note that to display
every level in a legend, the layer should use \code{show.legend = TRUE}.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{colours}{A character vector of colours used as key points in the colour
ramp that variables are mapped to. For details on how supplied colours are
used to construct the resulting palette, see \code{\link[=bivar_palette]{bivar_palette()}} and
\code{\link[=bivar_fade_palette]{bivar_fade_palette()}}.}

\item{palette_fun}{A palette function that, when called with \code{colours} and
\code{n_breaks}, returns a character vector of colours for all binned
combinations. If \code{NULL}, the default, \code{\link[=bivar_palette]{bivar_palette()}} is used.}

\item{palette_params}{A list of additional arguments passed to \code{palette}. For
details of arguments, see \code{\link[=bivar_palette]{bivar_palette()}} and \code{\link[=bivar_fade_palette]{bivar_fade_palette()}}.}

\item{n_breaks}{An integer or a length-two vector specifying the number of
bins for each variable. The default is 4 for both variables, and unequal
numbers of bins are supported.}

\item{bin_method}{A character string or a length-two vector specifying the
method used to bin each variable: \code{"equal"} (the default) or \code{"quantile"}.}

\item{var1_name, var2_name}{Optional names for \code{v1} and \code{v2}. Used as axis
titles in the legend. If \code{NULL}, the default, the names are taken from the
mapping.}

\item{super}{The super class to use for the constructed scale}
}
\value{
A \code{ScaleBivariate} ggproto object.
}
\description{
\code{bivariate_scale()} maps binned combinations of two variables to colour
dimensions and their combinations in perceptual colour space, supporting
the construction of bivariate choropleth maps.
}
\details{
It can be automatically dispatched in \code{aes()} using \code{duo()} and works
with any ggplot2 geom.
}
\examples{
# Basic bivariate map
ggplot(nc) +
  geom_sf(aes(fill = duo(value, sd)))

# Use an alternative bivariate palette
ggplot(nc) +
  geom_sf(aes(fill = duo(value, sd))) +
  scale_fill_bivariate(
    palette_fun = bivar_fade_palette,
    colours = c("#F6E8C3", "orange", "red")
  )

# Customize the number of bins
ggplot(nc) +
  geom_sf(aes(fill = duo(value, sd))) +
  scale_fill_bivariate(n_breaks = c(3, 4))
}
\seealso{
\link[ggplot2:Scale]{ggplot2::Scale} for the base ggproto class that all scale objects
inherit from.
}
\keyword{datasets}
