% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duo.R
\name{duo}
\alias{duo}
\alias{duo_exceed}
\title{Format input and assign the map class}
\usage{
duo(v1, v2)

duo_exceed(estimate, error)
}
\arguments{
\item{v1, v2}{Two variables to be used in constructing the map.}
}
\description{
\code{duo()} returns a list of sub-lists, each containing paired values of the two
variables, along with their variable names stored as the map class.
}
\examples{
data(nc)
value <- nc$value; sd <- nc$sd
res <- duo(value, sd)
class(res)
attr(res, "vars")

}
