% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duo.R
\name{duo}
\alias{duo}
\alias{duo_exceed}
\title{Format input and assign the map class}
\usage{
duo(v1, v2)

duo_exceed(estimate, error)
}
\arguments{
\item{v1, v2, estimate, error}{Input variables for constructing paired mappings.
\code{duo()} uses \code{v1} and \code{v2}, whereas \code{duo_exceed()} uses \code{estimate} and
\code{error} to represent the point and uncertainty measures, respectively.}
}
\value{
A list of sub-lists, each containing a pair of values from the two
variables, with attributes storing the variable names and the \code{"map"} class.
}
\description{
\code{duo()} and \code{duo_exceed()} create a paired mapping object that combines
two variables, records their names, and assigns the \code{"map"} class as an
attribute for use in aesthetic mappings.
}
\examples{
value <- nc$value
sd <- nc$sd

# Create a paired mapping object using duo()
res <- duo(value, sd)
res_exceed <- duo_exceed(value, sd)
class(res)
class(res_exceed)
attr(res, "vars")
attr(res_exceed, "vars")

}
