% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-bivariate.R
\docType{data}
\name{StatBivariate}
\alias{StatBivariate}
\alias{geom_sf_bivariate}
\title{Geom layer function for bivariate map on sf objects}
\format{
An object of class \code{StatBivariate} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatBivariate

geom_sf_bivariate(
  mapping = NULL,
  data = NULL,
  stat = StatBivariate,
  position = "identity",
  show.legend = TRUE,
  inherit.aes = TRUE,
  na.rm = FALSE,
  flip_axis = FALSE,
  breaks = c("quantile", "equal"),
  n_breaks = 4L,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping; must include \code{geometry}, \code{estimate}, and \code{error}.}

\item{data}{An sf object.}

\item{position}{Position adjustment.}

\item{show.legend}{Logical; whether to display a legend.}

\item{inherit.aes}{Logical; whether to inherit global aesthetics.}

\item{flip_axis}{Logical; whether to flip estimate/error axes.}

\item{breaks}{Whether to use "quantile" or "equal" bins.}

\item{n_breaks}{The number of breaks}

\item{...}{Additional arguments passed to \code{layer_sf()}.}

\item{geom}{Geom type to use.}
}
\value{
A ggproto object rendering the filled sf layer and a \code{coord_sf()}.
}
\description{
geom_sf_bivariate applies StatBivariate to sf data, classifying each region
into joint estimate/error bins and returning a map with bivariate colour scheme.
}
\examples{
set.seed(10086)

data(nc)

ggplot(nc) +
  geom_sf_bivariate(aes(primary = value, secondary = sd)) +
  scale_fill_bivariate(
    colrange = list(colour = c("gold", "red4"), difC = c(4, 4)),
    subtractive = FALSE,
    flip_vertical = FALSE,
    flip_horizontal = FALSE
  )

}
\keyword{datasets}
