% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-glyph.R
\name{geom_sf_glyph}
\alias{geom_sf_glyph}
\title{Glyph map}
\usage{
geom_sf_glyph(
  mapping = NULL,
  data = NULL,
  ...,
  shape = "circle",
  max_angle = NULL,
  size = 1,
  point_fun = sf::st_point_on_surface,
  border_colour = NA,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  angle_guide = TRUE,
  angle_name = waiver(),
  angle_order = 2
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{shape}{Glyph shape. One of \code{"circle"} (the default), \code{"square"},
\code{"triangle"}, \code{"hex"}, \code{"drop"}, or \code{"chernoff"}.}

\item{max_angle}{Maximum value of the \code{angle} aesthetic used for rescaling
glyph rotation.}

\item{size}{A positive numeric scaling factor controlling glyph size.}

\item{point_fun}{Function used to calculate the representative point for each
region. The default is usually \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}.}

\item{border_colour}{Colour used for glyph borders.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.

You can also set this to one of "polygon", "line", and "point" to
override the default legend.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{angle_guide}{Logical indicating whether to display a guide for the
\code{angle} aesthetic.}

\item{angle_name}{Title used for the angle guide.}

\item{angle_order}{Order of the angle guide relative to other guides.}
}
\value{
A list of ggplot2 layer objects.
}
\description{
\code{geom_sf_glyph()} generates a glyph map sf layer. A glyph map is a
centroid-based map, where each region is represented by a chosen glyph.
}
\details{
Regular shape glyphs can be treated as ordinary point-like symbols and used
together with \code{bivariate_scale()}. Drop-shaped glyphs use rotation angle to
represent uncertainty. Chernoff glyphs are adapted from the \code{ggChernoff}
package and work with the \code{smile} aesthetic.
}
\examples{
# Regular glyph map
ggplot(nc) +
  geom_sf_glyph(aes(colour = value), shape = "hex")

# Rotated drop glyph map
ggplot(nc) +
  geom_sf_glyph(
    aes(colour = value, angle = sd),
    shape = "drop"
  )

# Chernoff face glyph map
ggplot(nc) +
  geom_sf_glyph(
    aes(colour = value, smile = sd),
    shape = "chernoff"
  )
}
