% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-glyph.R
\name{geom_sf_glyph}
\alias{geom_sf_glyph}
\title{Geom layer function for glyph map on sf objects}
\usage{
geom_sf_glyph(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  size = 70,
  glyph = "icone",
  max_error = NULL,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping that must include \code{geometry}, \code{estimate}, and \code{error}.}

\item{data}{An sf object.}

\item{position}{Position adjustment.}

\item{show.legend}{Logical; whether to display a legend.}

\item{inherit.aes}{Logical; whether to inherit global aesthetics.}

\item{size}{Glyph scale factor.}

\item{glyph}{Glyph type: "icone" or "semi".}

\item{max_error}{Maximum error scale.}

\item{...}{Other params passed to layer()}

\item{geom}{Geom type, default "polygon"}
}
\description{
\code{geom_sf_glyph} applies \code{StatGlyph} to sf data and returns a centroid map
with color and rotation based on estimate and error values.
}
\examples{
set.seed(10086)

nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE) \%>\%
  mutate(value = rnorm(n()), sd = rnorm(n()))

ggplot(nc, aes(geometry = geometry, estimate = value, error = sd)) +
  geom_sf_glyph(size = 50, glyph = "icone") +
  coord_sf() + theme_void()

}
