% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-pixel.R
\name{geom_sf_pixel}
\alias{geom_sf_pixel}
\title{Geom layer function for pixel map on sf objects}
\usage{
geom_sf_pixel(
  mapping = NULL,
  data = NULL,
  pixel_size = 100,
  distribution = "uniform",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping that must include \code{geometry}, \code{estimate}, \code{error} and \code{id}.}

\item{data}{An sf object.}

\item{distribution}{Sampling distribution: "uniform", "normal", or "discrete".}

\item{show.legend}{Logical; show legend.}

\item{inherit.aes}{Logical; inherit global aesthetics.}

\item{...}{Other parameters passed to \code{layer()}.}

\item{pixelSize}{Number of grid cells along each axis.}

\item{id_col}{Name of the ID column.}
}
\value{
A polygon layer showing pixel values.
}
\description{
\code{geom_sf_pixel} applies \code{StatPixel} to sf data, pixelate each region and
sample each pixel's value from a specified distribution.
}
\examples{
data(nc)
ggplot() +
  geom_sf_pixel(
    data = nc,
    mapping = aes(geometry = geometry),
    id_col = "id",
    pixelSize = 40,
    distribution = "uniform"
  ) +
  scale_fill_distiller(palette = "Blues", direction = 1) +
  geom_sf(data = nc, fill = NA, color = 'black', size = 0.2)

}
