% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuzzy-package.R
\docType{package}
\name{ggincerta-package}
\alias{ggincerta}
\alias{ggincerta-package}
\title{ggincerta: Extending ggplot2 with specialized layers to simplify spatial uncertainty visualization.}
\description{
'ggfuzy' is an 'R' package designed to efficiently visualize spatial uncertainty by providing ggplot2 layer functions and corresponding scales, including bivariate maps, pixel maps, glyph maps, and exceedance probability maps.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
