% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggincerta-package.R
\docType{package}
\name{ggincerta-package}
\alias{ggincerta}
\alias{ggincerta-package}
\title{ggincerta: Extend 'ggplot2' with Layers and Scales for Spatial Uncertainty Visualization}
\description{
Provide specialized 'ggplot2' layers and scales for spatial uncertainty visualization, including bivariate choropleths, pixel maps, glyph maps, and exceedance probability maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maggiexma/ggincerta}
  \item Report bugs at \url{https://github.com/maggiexma/ggincerta/issues}
}

}
\author{
\strong{Maintainer}: Xueqi Ma \email{maggiexma07@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [thesis advisor]
}

Other contributors:
\itemize{
  \item Weihao Li [thesis advisor]
  \item Quan Vu [thesis advisor]
  \item Francis Hui [thesis advisor]
}

}
\keyword{internal}
