% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggincerta-package.R
\docType{package}
\name{ggincerta-package}
\alias{ggincerta}
\alias{ggincerta-package}
\title{ggincerta: Extending ggplot2 with specialized layers to simplify spatial uncertainty visualization.}
\description{
This package designed to efficiently visualize spatial uncertainty by providing ggplot2 layer functions and corresponding scales, including bivariate maps, pixel maps, glyph maps, and exceedance probability maps.
}
\author{
\strong{Maintainer}: Xueqi Ma \email{first.last@example.com} [copyright holder]

Authors:
\itemize{
  \item Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [thesis advisor]
}

Other contributors:
\itemize{
  \item Patrick Li [thesis advisor]
  \item Quan Vo [thesis advisor]
  \item Francis Hui [thesis advisor]
}

}
\keyword{internal}
