% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-glyph.R
\docType{data}
\name{GuideGlyph}
\alias{GuideGlyph}
\alias{guide_glyph}
\title{Drop glyph guide}
\format{
An object of class \code{GuideGlyph} (inherits from \code{Guide}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GuideGlyph

guide_glyph(
  theme = NULL,
  title = waiver(),
  order = 99,
  position = NULL,
  direction = "vertical"
)
}
\arguments{
\item{theme}{A \code{\link[ggplot2:theme]{theme}} object to style the guide individually or
differently from the plot's theme settings. The \code{theme} argument in the
guide partially overrides, and is combined with, the plot's theme.
Arguments that apply to a single legend are respected, most of which have
the \code{legend}-prefix. Arguments that apply to combined legends
(the legend box) are ignored, including \code{legend.position},
\verb{legend.justification.*}, \code{legend.location} and \verb{legend.box.*}.}

\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{order}{positive integer less than 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{position}{A character string indicating where the legend should be
placed relative to the plot panels.
One of "top", "right", "bottom", "left", or "inside".}

\item{direction}{A character string indicating the direction of the guide.
One of "horizontal" or "vertical".}
}
\description{
Guide for drop glyph consists of two axes and a sequence of example drop
grobs illustrating the mapping between uncertainty and glyph rotation angles.
}
\details{
Rotation angles range from \eqn{\pi} to \eqn{-\pi}, corresponding to the
minimum and maximum uncertainty values respectively.
}
\keyword{datasets}
