% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-bivariate-manual.R
\name{manual_bivariate_scale}
\alias{manual_bivariate_scale}
\alias{scale_fill_bivariate_manual}
\alias{scale_colour_bivariate_manual}
\alias{scale_color_bivariate_manual}
\title{Create your own bivariate colour scale}
\usage{
manual_bivariate_scale(
  aesthetics,
  ...,
  values,
  name = waiver(),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  na.value = NA,
  na.translate = TRUE,
  drop = FALSE,
  guide = guide_bivariate(),
  n_breaks = c(4, 4),
  bin_method = c("equal", "equal"),
  var1_name = NULL,
  var2_name = NULL,
  super = ScaleBivariate
)

scale_fill_bivariate_manual(
  ...,
  values,
  name = waiver(),
  var1_name = NULL,
  var2_name = NULL,
  n_breaks = c(4, 4),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  bin_method = c("equal", "equal"),
  na.value = NA,
  na.translate = TRUE,
  aesthetics = "fill",
  guide = guide_bivariate()
)

scale_colour_bivariate_manual(
  ...,
  values,
  name = waiver(),
  var1_name = NULL,
  var2_name = NULL,
  n_breaks = c(4, 4),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  bin_method = c("equal", "equal"),
  na.value = NA,
  na.translate = TRUE,
  aesthetics = "colour",
  guide = guide_bivariate()
)

scale_color_bivariate_manual(
  ...,
  values,
  name = waiver(),
  var1_name = NULL,
  var2_name = NULL,
  n_breaks = c(4, 4),
  breaks = list(waiver(), waiver()),
  labels = list(waiver(), waiver()),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  bin_method = c("equal", "equal"),
  na.value = NA,
  na.translate = TRUE,
  aesthetics = "colour",
  guide = guide_bivariate()
)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{...}{Other arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}

\item{values}{A character vector of colours. The length should be at least
the product of \code{n_breaks}, giving one colour for each combination of the
two binned variables.}

\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{A list of two numeric vectors specifying bin boundaries for
each variable. If \code{waiver()}, breaks are computed from the data according
to \code{n_breaks} and \code{bin_method}.}

\item{labels}{A list of two character vectors or labelling functions
used to label the bin boundaries for each variable. If \code{waiver()},
default numeric labels are used.}

\item{limits}{A list of two numeric vectors specifying the range of
values to include for each variable.}

\item{transform}{A list of two transformations applied to the variables
before binning. Each element can be a transformation name or a
transformer object accepted by \code{\link[scales:as.transform]{scales::as.transform()}}.}

\item{na.value}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}

\item{na.translate}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}

\item{drop}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} includes the levels in the factor. Please note that to display
every level in a legend, the layer should use \code{show.legend = TRUE}.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{n_breaks}{An integer or a length-two vector specifying the number of
bins for each variable. The default is 4 for both variables, and unequal
numbers of bins are supported.}

\item{bin_method}{A character string or a length-two vector specifying the
method used to bin each variable: \code{"equal"} (the default) or \code{"quantile"}.}

\item{var1_name, var2_name}{Optional names for \code{v1} and \code{v2}. Used as axis
titles in the legend. If \code{NULL}, the default, the names are taken from the
mapping.}

\item{super}{The super class to use for the constructed scale}
}
\description{
These scales allow users to provide all colours used for the combinations of
two variables in a bivariate colour scale.
}
\examples{
ggplot(nc, aes(fill = duo(value, sd))) +
  geom_sf() +
  scale_fill_bivariate_manual(
    values = c(
      "#F7F4F9", "#D4B9DA", "#C994C7", "#980043",
      "#E0ECF4", "#BFD3E6", "#9EBCDA", "#8856A7",
      "#D0D1E6", "#A6BDDB", "#74A9CF", "#2B8CBE",
      "#B8E186", "#7FBC41", "#4D9221", "#276419"
    )
  )
}
