% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-vsup.R
\name{ScaleVSUP}
\alias{ScaleVSUP}
\alias{scale_fill_vsup}
\alias{scale_colour_vsup}
\title{Value-Suppressing Uncertainty Palette (VSUP) scale}
\usage{
ScaleVSUP

scale_fill_vsup(
  name = waiver(),
  colours = RColorBrewer::brewer.pal(3, "YlGnBu"),
  layers = 4,
  branch = 2L,
  breaks = list(NULL, NULL),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  title_value = "Value",
  title_uncertainty = "Uncertainty",
  na.value = NA,
  na.translate = TRUE,
  aesthetics = "fill",
  max_light = 0.7,
  max_desat = 0.9,
  pow_light = 1,
  pow_desat = 1,
  space = "Lab",
  guide = guide_vsup(),
  ...
)

scale_colour_vsup(
  name = waiver(),
  colours = RColorBrewer::brewer.pal(3, "YlGnBu"),
  layers = 4,
  branch = 2L,
  breaks = list(NULL, NULL),
  limits = list(NULL, NULL),
  transform = list("identity", "identity"),
  title_value = "Value",
  title_uncertainty = "Uncertainty",
  na.value = NA,
  na.translate = TRUE,
  aesthetics = "colour",
  max_light = 0.7,
  max_desat = 0.9,
  pow_light = 1,
  pow_desat = 1,
  space = "Lab",
  guide = guide_vsup(),
  ...
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{colours}{A character vector of colours used as key points in the value
colour scale. See \code{\link[=vsup_palette]{vsup_palette()}} for details.}

\item{layers}{An integer specifying the number of uncertainty levels.}

\item{branch}{An integer specifying the branching factor used to allocate
value bins across uncertainty levels. The maximum number of value bins is
\code{branch^(layers - 1)}, and higher uncertainty levels are assigned fewer
value bins.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{transform}{A list of two transformations applied to the variables
before binning. Each element can be a transformation name or a
transformer object accepted by \code{\link[scales:as.transform]{scales::as.transform()}}.}

\item{title_value, title_uncertainty}{Optional titles for the value and
uncertainty dimensions in the guide.}

\item{na.value}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}

\item{na.translate}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}

\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{max_light}{A numeric value specifying the maximum amount of lightening
applied across uncertainty levels.}

\item{max_desat}{A numeric value specifying the maximum amount of
desaturation applied across uncertainty levels.}

\item{pow_light, pow_desat}{Numeric values controlling the rate of lightening
and desaturation across uncertainty levels.}

\item{space}{A character string specifying the colour space used for colour
interpolation.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{...}{Other arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
This scale implements Value-Suppressing Uncertainty Palettes (VSUPs),
proposed by Correll et al. (2018). The main idea is to suppress colour
variation in regions with higher uncertainty, thereby directing visual
attention towards more reliable value differences.
}
\seealso{
Correll et al. (2018) \doi{10.1145/3173574.3174216} for technical
details.
}
